Add func to switch between Scheme file and interp
This commit adds a couple of procedures to switch directly to the
window with scheme interpreter's buffer and to switch back to
latest scheme file from interpreter's buffer. Both switching to
and fro have the same kbd so now it's pretty comfortable to move
between file editing and checking results (similar to ghc's
comfort).